+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Newlines are no longer semi-randomly printed to standard error in preprocessing and assembling mode.
 
 ### Added
-- The `--credits` flags.
+- The `--throttle` flag, to limit the number of operations executed on a single run.
+- The `--credits` flag.
 
 ## [0.7.0] - 2019-02-13
 ### Added